requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 
openai==1.1.0
langchain==0.2.11
langchain-community>=0.0.1
